+++ Changelog.md
@@ -2,6 +2,16 @@
 
 This page is a list of _notable_ changes made in each version.
 
+## v0.2.0 - Feb 13, 2015
+
+Features:
+
+- Added the `-CalculateDiskSpaceReclaimed` switch to show how much disk space was reclaimed in the output.
+
+BREAKING CHNAGES:
+
+- Renamed the `Clean-GitRepositories` cmdlet to `Invoke-GitClean`.
+
 ## v0.1.0 - Feb 12, 2015
 
 - Initial release